+++ src/queue-add-job.js
@@ -3,6 +3,7 @@ const Promise = require('bluebird')
 const enums = require('./enums')
 const queueProcess = require('./queue-process')
 const dbResult = require('./db-result')
+const jobAddLog = require('./job-add-log')
 const jobParse = require('./job-parse')
 
 module.exports = function queueAddJob (q, job, skipStatusCheck) {
@@ -14,7 +15,8 @@ module.exports = function queueAddJob (q, job, skipStatusCheck) {
       return Promise.reject(new Error(enums.message.jobAlreadyAdded))
     }
     if (!skipStatusCheck) { oneJob.status = enums.status.waiting }
-    const log = oneJob.createLog(
+    const log = jobAddLog.createLogObject(oneJob,
+      null,
       enums.message.jobAdded,
       enums.log.information,
       enums.status.waiting)